Makefile: do not override RELEASE when build with XML
SCENARIO XML file has included RELEASE or DEBUG info already, so if RELEASE
is not specified in make command, Makefile should not override RELEASE info
in SCENARIO XML. If RELEASE is specified in make command, then RELEASE info
in SCENARIO XML could be overridden by make command.

The patch also fixed a issue that get correct board defconfig when build
hypervisor from TARGET_DIR;

Tracked-On: #4688

+++ hypervisor/scripts/makefile/kconfig.mk
@@ -38,7 +38,9 @@ ifeq ($(shell [ $(HV_OBJDIR)/$(HV_CONFIG) -nt $(HV_OBJDIR)/$(HV_CONFIG_MK) ] && 
 endif
 $(eval $(call override_config,BOARD,apl-mrb))
 $(eval $(call override_config,SCENARIO,))
-$(eval $(call override_config,RELEASE,n))
+$(eval $(call override_config,RELEASE,y))
+
+override RELEASE := $(CONFIG_RELEASE)
 
 $(eval $(call check_dep_exec,python3,KCONFIG_DEPS))
 $(eval $(call check_dep_py3lib,kconfiglib,KCONFIG_DEPS))
@@ -72,7 +74,10 @@ defconfig: $(KCONFIG_DEPS)
 		if [ "$(KCONFIG_FILE)" != "" ] && [ -f $(KCONFIG_FILE) ]; then \
 			echo "Writing $(HV_OBJDIR)/$(HV_CONFIG) with $(KCONFIG_FILE)"; \
 			cp $(KCONFIG_FILE) $(HV_OBJDIR)/$(HV_CONFIG); \
-		elif [ "$(CONFIG_XML_ENABLED)" = "true" ] && [ "$(TARGET_DIR)" != "" ] && [ -d $(TARGET_DIR) ]; then \
+		elif [ "$(TARGET_DIR)" != "" ] && [ -d $(TARGET_DIR) ]; then \
+			if [ ! -f $(TARGET_DIR)/$(BOARD).config ]; then \
+				echo "Board defconfig file $(BOARD).config is not found under $(TARGET_DIR)."; exit 1; \
+			fi; \
 			echo "Writing $(HV_OBJDIR)/$(HV_CONFIG) with $(TARGET_DIR)/$(BOARD).config"; \
 			cp $(TARGET_DIR)/$(BOARD).config $(HV_OBJDIR)/$(HV_CONFIG); \
 		fi; \
